pyproject.toml
[tool.poetry]
name = "zxcs_db"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "0.23.2"
loguru = "0.7.2"
fastapi = "0.104.0"
sqlalchemy = "^2.0.22"
redis = "^5.0.1"
arrow = "^1.3.0"
pydantic-settings = "^2.0.3"
pyyaml = "^6.0.1"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
